test(linter): move issue 2136 regressions to rule test macros
Replace the standalone integration test harness with the existing rule test macros in no-fully-qualified-global-class-like.
